float all over

+++ src/data/dataset.py
@@ -1,4 +1,4 @@
-from numpy import dtype, float64
+from numpy import dtype, float32, float64
 import torch
 from torch.utils.data import Dataset
 from sklearn.preprocessing import OneHotEncoder
@@ -109,7 +109,7 @@ class EmbeddingVector(Dataset):
             ]
         )
         
-        input_embed = torch.tensor(np.concatenate([encoding, bio_encoding], dtype=float))
+        input_embed = torch.tensor(np.concatenate([encoding, bio_encoding], dtype=float32))
         # print(input_embed)
         return {
             # "input_embed": torch.tensor(np.concatenate([encoding, bio_encoding, tag_encoding, other_encoding.astype("float64")], dtype=float64))
